requirements.txt
asgiref==3.5.2
Babel==2.10.3
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Django==4.1
django-allauth==0.51.0
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-money==3.0.0
django-phonenumber-field==6.3.0
django-taggit==3.0.0
idna==3.3
oauthlib==3.2.0
phonenumbers==8.12.53
Pillow==9.2.0
py-moneyed==2.0
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
typing-extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
